+++ src/main/java/com/devbliss/gpullr/controller/dto/RankingListDto.java
@@ -1,0 +1,13 @@
+package com.devbliss.gpullr.controller.dto;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class RankingListDto {
+
+  public List<RankingDto> rankings = new ArrayList<>();
+
+  public String calculationDate;
+
+  public String rankingScope;
+}