+++ changelog.md
@@ -2,6 +2,7 @@
 ### Version 0.1.0-SNAPSHOT
 
 
+ - [bug] fixed ranking calculation / close dates
  - [bug] fixed date time formatting for pullrequest dto
  - [bug] first attempt to solve fetch problems
  - [bug] pull requests w/o user session works again (with default sort order)
@@ -29,5 +30,5 @@
  - [feature] project setup
  - [initial] initial commit
 
--- Last change by: Philipp Karstedt <philipp.karstedt@devbliss.com> Mon Mar 16 17:30:28 CET 2015
+-- Last change by: hschuetz <henning.schuetz@devbliss.com> Wed Mar 18 13:16:46 CET 2015